requirements.txt
asgiref==3.2.7
boto3==1.12.40
botocore==1.15.40
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-admin-sortable==2.2.3
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-filter==2.2.0
django-fsm==2.7.0
django-fsm-admin==1.2.4
django-import-export==2.0.2
django-ipware==2.1.0
django-reversion==3.0.7
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
drf-recaptcha==2.0.0
et-xmlfile==1.0.1
google-api-core==1.17.0
google-auth==1.14.0
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==2.0.0
google-resumable-media==0.5.0
google-api-python-client==1.8.2
google-auth-httplib2==0.0.3
google-auth-oauthlib--0.4.1
googleapis-common-protos==1.51.0
googledrivedownloader==0.4
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
jmespath==0.9.5
MarkupPy==1.14
msgpack==1.0.0
odfpy==1.4.1
openpyxl==3.0.3
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
